Added authorization and authentication, also reformatted code a DB a little bit
