+++ movie_app/alembic/versions/5bbde2a096b9_roles_in_user.py
@@ -1,0 +1,28 @@
+"""roles_in_user
+
+Revision ID: 5bbde2a096b9
+Revises: 30a7551ba775
+Create Date: 2020-12-22 12:52:01.914021
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '5bbde2a096b9'
+down_revision = '30a7551ba775'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('user', sa.Column('role', sa.String(length=64), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('user', 'role')
+    # ### end Alembic commands ###
